requirements.txt
pandas==2.2.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.4
numpy
Pillow
scikit-learn
fastapi
uvicorn
python-multipart
jinja2
opencv-python
pyexiftool